# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: URL Shortener API
  description: A simple URL shortener service
  version: 1.0.0

tags:
  - name: url-shortener
    description: URL shortening operations

paths:
  /shorten:
    post:
      tags:
        - url-shortener
      operationId: shortenUrl
      summary: Shorten a URL
      description: Creates a shortened URL for the given original URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortenRequest'
      responses:
        '201':
          description: URL shortened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortenResponse'
        '400':
          description: Invalid URL format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Code generation failed after maximum retry attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{code}:
    get:
      tags:
        - url-shortener
      operationId: redirectToOriginalUrl
      summary: Redirect to original URL
      description: Redirects to the original URL associated with the given short code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to original URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{code}/info:
    get:
      tags:
        - url-shortener
      operationId: getUrlInfo
      summary: Get URL info
      description: Returns metadata about a shortened URL
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: URL info retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlInfoResponse'
        '404':
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ShortenRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The original URL to shorten
          example: "https://www.originenergy.com.au/electricity-gas/plans.html"

    ShortenResponse:
      type: object
      required:
        - shortUrl
        - originalUrl
      properties:
        shortUrl:
          type: string
          description: The generated short URL
          example: "http://localhost:8080/a1B2c3"
        originalUrl:
          type: string
          description: The original URL
          example: "https://www.originenergy.com.au/electricity-gas/plans.html"

    UrlInfoResponse:
      type: object
      required:
        - code
        - originalUrl
        - createdAt
      properties:
        code:
          type: string
          description: The short code
          example: "a1B2c3"
        originalUrl:
          type: string
          description: The original URL
          example: "https://www.originenergy.com.au/electricity-gas/plans.html"
        createdAt:
          type: string
          format: date-time
          description: When the URL was shortened
          example: "2026-03-26T10:30:00Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "URL not found"